+++ contracts/Exchange.sol
@@ -1,4 +1,4 @@
-pragma solidity ^0.4.21;
+pragma solidity ^0.4.23;
 
 import "./ExchangeInterface.sol";
 import "./Libraries/OrderLibrary.sol";
@@ -17,7 +17,7 @@ contract Exchange is Ownable, ExchangeInterface {
 
     ExchangeLibrary.Exchange public exchange;
 
-    function Exchange(uint _takerFee, address _feeAccount, VaultInterface _vault) public {
+    constructor(uint _takerFee, address _feeAccount, VaultInterface _vault) public {
         require(address(_vault) != 0x0);
         setFees(_takerFee);
         setFeeAccount(_feeAccount);