+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
  - Minor cleanup to trade function
 
+### Changed
+ - Deduplicated withdraw handling.
+
 ## [2.0.0] - 2018-04-18
 
 ### Changed