+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
  - Deduplicated withdraw handling.
 
+### Fixed
+ - Fixes ERC777 implementation
+
 ## [2.0.0] - 2018-04-18
 
 ### Changed