+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
  - Solidity version 0.4.23
+ -  Using new [Signature Validator](https://github.com/DexyProject/SignatureValidator) package.
 
 ## [2.1.0] - 2018-04-26