+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [unreleased]
 
 ### Changed
- -  Using new [Signature Validator](https://github.com/DexyProject/SignatureValidator) package.
+ - Solidity version 0.4.23
+ - Using new [Signature Validator](https://github.com/DexyProject/SignatureValidator) package.
 
 ## [2.1.0] - 2018-04-26